Provide nicer error messaging for unknown keys
The error message when a new, unknown racial group was added was broken (it didn't correctly place the new group name in the message). This fixes that, but goes a step farther by creating a full list of the expected and actual groups and printing what was different. Now we have more informative errors like:

    FormatError: Racial groups are different -- removed: {'Multi-Race', 'Asian / Pacific Islander, non-Hispanic', 'Other', 'Black / African American, non-Hispanic'}, added: {'Asian, non-Hispanic', 'Black/African American, non-Hispanic', 'Native Hawaiian and other Pacific Islander, non-Hispanic', 'American Indian/Alaska Native, non-Hispanic', 'Multi-racial, non-Hispanic'}

(Even though they may get long, these are hugely useful.)